setting up and show debugtoolkit in coordinator
